requirements.txt
torch
casadi
mpctools
matplotlib
gym
tensorboard
pyyaml
numpy==1.19.1